Big refactor.

+++ src/api.rs
@@ -67,7 +67,7 @@ pub struct TimeValue {
     pub ts: chrono::DateTime<chrono::Utc>,
     #[serde(skip_serializing_if = "Option::is_none")]
     #[serde(rename = "ms")]
-    pub latency_ms: Option<i32>,
+    pub latency_ms: Option<u64>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub err: Option<Error>,
 }
